Make use of the fact that C# is OO and we can pass around refernces
Instead of using the previous approach where separate helper methods were
implemented to check whether a location/enemy/item/attack ("World stuff")
exists, what its name or ID, we will now use a more sophisticated method.

A new class was implemented which contains the elements and their ID and
string key and exposes the previous methods, supporting access with ID
and name.

This also eliminates the double lists and ID-name conversions previously
implemented.

NOTE: This commit does NOT build, and that is intended. Lots of code need
to be aligned to this new approach, as the "external" API of class World
broke.