Added null checks for referenced entities
Currently, the program will crash or open the debugger when a null
reference is found, which is not bad while developing.
For the release it would be a good idea to create a custom exception for
this situation and throw that. The method building the world can then
catch it and display a more helpful error message (like which entity in
which mod is erroneous).